refactor(files): Create separated files for each endpoint
