--- src/controllers/_file_controllers.py
@@ -1,78 +1,0 @@
-import json
-from flask import request
-from src.config.soap_client import soap_client
-
-
-def check_state_handler(token, file_uuid):
-    try:
-        response = soap_client.service.file_check(
-            {"token": token, "fileUUID": file_uuid}
-        )
-
-        has_success_code = str(response.code).startswith("20")
-        if not has_success_code:
-            return {"msg": response.msg}, response.code
-
-        return {
-            "msg": "File status has been obtained successfully",
-            "ready": response.ready,
-        }, response.code
-    except Exception as e:
-        print("[Exception] check_state_handler ->", str(e))
-        return {"msg": "There was an error checking the file state"}, 500
-
-
-def rename_handler(token, file_uuid):
-    try:
-        data = json.loads(request.data)
-        new_name = data["newName"]
-        if not new_name or len(new_name) == 0:
-            return {"msg": "New name is required"}, 400
-
-        request_data = {"token": token, "fileUUID": file_uuid, "newName": new_name}
-        response = soap_client.service.file_rename(request_data)
-
-        if response["error"] is True:
-            return {"msg": response["msg"]}, response["code"]
-        else:
-            return {"msg": "The file has been renamed successfully"}, 200
-
-    except ValueError:
-        return {"msg": "Invalid JSON data provided in the request"}, 400
-
-    except Exception:
-        return {"msg": "There was an error renaming the file"}, 500
-
-
-def create_new_dir_handler(token):
-    try:
-        data = json.loads(request.data)
-
-        directoryName = data.get("directoryName")
-        location = data.get("location")
-
-        if not directoryName or not location:
-            return {"msg": "Required fields are missing in JSON data"}, 400
-
-        response = soap_client.service.file_new_dir(
-            {
-                "directoryName": directoryName,
-                "location": location,
-                "token": token,
-            }
-        )
-
-        if response["error"] is True:
-            return {"msg": response["msg"]}, response["code"]
-        else:
-            return {
-                "msg": "New directory created successfully",
-                "directoryUUID": response["fileUUID"],
-            }, response["code"]
-
-    except ValueError:
-        return {"msg": "Invalid JSON data provided in the request"}, 400
-
-    except Exception as e:
-        print("[Exception] create_new_directory ->", str(e))
-        return {"msg": "There was an error creating the new directory"}, 500
